wtex.sh, doc/T@x.tex: como wpdf.sh
Cambio wtex.sh como wpdf.sh,
y actualizo la documentación correspondiente.

+++ exe/sh/wtex.sh
@@ -12,6 +12,7 @@ if test "$APATH" != "$(pwd)" ; then
 fi
 
 TEXFILE=$(basename "$1")
+BASEFILE=$(basename "$1" .tex)
 MFFILE="auxiliar.mf"
 INDFILE="auxiliar.ind"
 INTFILE="auxiliar.int"
@@ -19,34 +20,44 @@ NDXFILE="auxiliar.ndx"
 ABCFILE="auxiliar.abc"
 AUXFILE="auxiliar.aux"
 
-doindex() {
-if test -e $INDFILE ; then
-   echo "readtex < $INDFILE > $INTFILE"
-         readtex < $INDFILE > $INTFILE
-   echo "reindex \"$TEXFILE\""
-         reindex "$TEXFILE"
-   if test -e $NDXFILE ; then
-      echo "index $NDXFILE < $INTFILE > $ABCFILE"
-            index $NDXFILE < $INTFILE > $ABCFILE
-   else
-      echo "texsort < $INTFILE > $ABCFILE"
-            texsort < $INTFILE > $ABCFILE
-   fi
+if test -f "$BASEFILE.ndx" ; then
+   echo "Pass 0"
+   echo "iconv -f UTF-8 -t ISO-8859-1 -o $NDXFILE \"$BASEFILE.ndx\""
+   iconv -f UTF-8 -t ISO-8859-1 -o $NDXFILE "$BASEFILE.ndx"
 fi
-}
 
-if test -e $MFFILE ; then
-   echo "Only one pass!"
-   doindex
+PREMF=""
+PREAUX=""
+PREIND=""
+POSTMF=$(md5sum $MFFILE 2>&1)
+POSTAUX=$(md5sum $AUXFILE 2>&1)
+POSTIND=$(md5sum $INDFILE 2>&1)
+i=0
+until [ "$POSTMF" = "$PREMF" -a "$POSTAUX" = "$PREAUX" -a "$POSTIND" = "$PREIND" ]
+do
+   i=$(expr $i + 1)
+   echo "Pass $i"
+   if test -e $INDFILE -a "$POSTIND" != "$PREIND" ; then
+      echo "readtex < $INDFILE > $INTFILE"
+            readtex < $INDFILE > $INTFILE
+      if test -e $NDXFILE ; then
+         echo "index $NDXFILE < $INTFILE > $ABCFILE"
+               index $NDXFILE < $INTFILE > $ABCFILE
+      else
+         echo "texsort < $INTFILE > $ABCFILE"
+               texsort < $INTFILE > $ABCFILE
+      fi
+   fi
    echo "tex '&spplain' \"$TEXFILE\""
+   PREMF="$POSTMF"
+   PREAUX="$POSTAUX"
+   PREIND="$POSTIND"
    tex '&spplain' "$TEXFILE"
-else
-   echo "First pass"
-   doindex
-   echo "tex '&spplain' \"$TEXFILE\""
-   tex '&spplain' "$TEXFILE"
-   if test -e $MFFILE ; then
-      echo "Second pass"
+   POSTMF=$(md5sum $MFFILE 2>&1)
+   POSTAUX=$(md5sum $AUXFILE 2>&1)
+   POSTIND=$(md5sum $INDFILE 2>&1)
+   if test "$POSTMF" != "$PREMF" ; then
+      echo "mf $MFFILE"
       mf $MFFILE
       LOGFILE="auxiliar.log"
       GFNAME=$(awk '/^Output written on / {print $4}' $LOGFILE)
@@ -54,11 +65,8 @@ else
       PKFILE=${GFFILE%gf}pk
       echo "gftopk $GFFILE $PKFILE"
       gftopk $GFFILE $PKFILE
-      doindex
-      echo "tex '&spplain' \"$TEXFILE\""
-      tex '&spplain' "$TEXFILE"
    fi
-fi
+done
 
 exit
